Function move in progress.
